Update Tipe Data

+++ kopiku/app/src/main/java/com/ppbl2021_rpl_kaizen/kopiku/ui/Dashboard.kt
@@ -57,7 +57,7 @@ class Dashboard : AppCompatActivity() {
                     for (document in result) {
                         val id = document.id
                         val namaBarang = document.get("nama_barang").toString()
-                        val hargaBarang = document.get("harga_barang").toString()
+                        val hargaBarang = document.get("harga_barang").toString().toInt()
                         val kategori = document.get("kategori").toString()
                         quotesList.add(Coffee(id, namaBarang , hargaBarang, kategori))
                     }
@@ -99,3 +99,4 @@ class Dashboard : AppCompatActivity() {
         }
     }
 }
+
